+++ CHANGELOG.md
@@ -6,8 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+
+## [1.4.15](https://github.com/kb-dk/kb-util/tree/kb-util-1.4.15)
+### Added
 - `XMLStepper.serializeSubElements` for streaming serialization (aka extraction) of XML sub elements.
 - Extension of the Timing class with several helper methods for lambda and streaming
+- Make HttpServletResponse optional in wrap for StreamingExportFactory
 
 ## [1.4.14](https://github.com/kb-dk/kb-util/tree/kb-util-1.4.14)
 ### Bugfix